modified css file
The probleem of hover effect on the items is solved. 
now when you hover on aan item there will be no border on the bottom, but just the title will have underline decoration. 
The challenge was : hovering on the item and just one element in it should have underline decoration. In this case it is the heading, while the rest of the content has no underline.
please: this point cost me lost of thinking. it is simple to doe , but difficult to figure it out . so would you please have a look carfuly on the original site and also on the work of other members of class 22.